extract functionality to EdgeMarker class
